refactor: fix embasy time driver implementation
The original time driver is actually running on so-called "low-power" mode.
i.e. it does not tick. It sleeps(WFI) to next alarm instead.
The Embassy timer driver implies the tick should be activated every TICK_HZ.
This is OK for EXTI awaits, which also awakes the core.But it'll fail when
using peripherals like DMA(requires CPU time to finish job).