feat(otgfs): Implement OTG_FS USB Driver
The primary purpose of this change is to introduce an initial
implementation of the USB OTG FS device driver. Although this should
also applie to other SKUs that have the USB FS (without OTG) but those
are currently untested due to the lack of said hardware for testing.

In the process of implementing support for USB, we noticed that the WFI
instruction in the Qingke IP has significant flaws causing the USB
peripheral unable to access memory correctly while the CPU is under WFI.
The exact reason behind this is unclear, but a workaround has been put
inplace by using a new executor we added to embassy named `arch-spin`
which avoids the use of WFI instruction altogether.

Also there were several updates to the peripheral access crate (PAC) to
ensure the USB driver code is as readable as possible. The update may
have not been applied to all device families, especially those using
USB_FS instead of having the full OTG peripheral.

The code has been tested on CH32V307 and CH32V305 with the following
applications:
- USB DFU: mostly to exercise usb control transfer
- USB HID: mostly to exercise usb interrupt transfer

We will also provide a simple USB DFU example in the `examples/ch32v307`
directory for those who want to just quickly verify any functionalities.

Co-authored-by: Dummyc0m <xieyuanchu@gmail.com>
Co-authored-by: Codetector <codetector@codetector.org>